Fix Authorize

+++ src/main/java/ru/skypro/homework/service/impl/CommentServiceImpl.java
@@ -1,5 +1,10 @@
 package ru.skypro.homework.service.impl;
 
+import org.springframework.security.access.AccessDeniedException;
+import org.springframework.security.access.prepost.PostAuthorize;
+import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 import org.springframework.transaction.annotation.Transactional;
@@ -7,6 +12,7 @@ import ru.skypro.homework.dto.Comment;
 import ru.skypro.homework.dto.Comments;
 import ru.skypro.homework.dto.CreateOrUpdateComment;
 
+import ru.skypro.homework.dto.Role;
 import ru.skypro.homework.model.AdEntity;
 import ru.skypro.homework.model.CommentEntity;
 import ru.skypro.homework.repo.AdRepository;
@@ -14,6 +20,7 @@ import ru.skypro.homework.repo.CommentRepository;
 import ru.skypro.homework.repo.UserRepo;
 import ru.skypro.homework.service.CommentMapper;
 import ru.skypro.homework.service.CommentService;
+import ru.skypro.homework.util.exceptions.NotFoundException;
 
 import java.time.LocalDateTime;
 import java.time.ZoneOffset;
@@ -59,10 +66,15 @@ public class CommentServiceImpl implements CommentService {
     @Override
     @Transactional
     public void deleteComment(Integer adId, Integer commentId) {
+        CommentEntity comment = commentRepository.findById(commentId).orElseThrow(() -> new NotFoundException("Комментарий с ID: "+ commentId+" не найден"));
+        if(comment.getAuthor().getRole().equals(Role.ADMIN) || comment.getAuthor().getLogin().equals(getMe())){
         commentRepository.deleteByAdIdAndId(adId,commentId);
+        }
+        else throw new AccessDeniedException("Нет доступа");
     }
 
     @Override
+    @PostAuthorize("returnObject.author == principal.username or hasRole('ADMIN')")
     public Comment updateComment(Integer adId, Integer commentId, CreateOrUpdateComment createOrUpdateComment) {
         LocalDateTime now = LocalDateTime.now();
         long milliseconds = now.toInstant(ZoneOffset.UTC).toEpochMilli();
@@ -72,6 +84,11 @@ public class CommentServiceImpl implements CommentService {
         commentRepository.save(commentEntity);
         Comment comment = commentMapper.commentEntityToCommentDTO(commentEntity);
         return comment;
+    }
+
+    private String getMe(){
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        return authentication.getName();
     }
 
 }
